+++ src/llms/chatgpt.ts
@@ -2,8 +2,9 @@
 
 import { env } from 'node:process'
 import { OpenAI } from 'openai'
-import { GPT_MODELS } from '../utils/globals/llms'
-import { err, logLLMCost } from '../utils/logging'
+import { GPT_MODELS } from '../utils/llm-utils'
+import { err } from '../utils/logging'
+import { logLLMCost } from '../utils/llm-utils'
 import type { ChatGPTModelType } from '../utils/types/llms'
 
 /**
