+++ src/llms/deepseek.ts
@@ -7,8 +7,9 @@
 
 import { env } from 'node:process'
 import { OpenAI } from 'openai'
-import { err, logLLMCost } from '../utils/logging'
-import { DEEPSEEK_MODELS } from '../utils/globals/llms'
+import { err } from '../utils/logging'
+import { logLLMCost } from '../utils/llm-utils'
+import { DEEPSEEK_MODELS } from '../utils/llm-utils'
 import type { DeepSeekModelType } from '../utils/types/llms'
 
 /**
