+++ src/llms/gemini.ts
@@ -2,8 +2,9 @@
 
 import { env } from 'node:process'
 import { GoogleGenerativeAI } from "@google/generative-ai"
-import { GEMINI_MODELS } from '../utils/globals/llms'
-import { err, logLLMCost } from '../utils/logging'
+import { GEMINI_MODELS } from '../utils/llm-utils'
+import { err } from '../utils/logging'
+import { logLLMCost } from '../utils/llm-utils'
 import type { GeminiModelType } from '../utils/types/llms'
 
 /**
