+++ src/llms/groq.ts
@@ -1,8 +1,9 @@
 // src/llms/groq.ts
 
 import { env } from 'node:process'
-import { GROQ_MODELS } from '../utils/globals/llms'
-import { err, logLLMCost } from '../utils/logging'
+import { GROQ_MODELS } from '../utils/llm-utils'
+import { err } from '../utils/logging'
+import { logLLMCost } from '../utils/llm-utils'
 import type { GroqModelType, GroqChatCompletionResponse } from '../utils/types/llms'
 
 /**
