+++ src/process-commands/rss.ts
@@ -19,9 +19,8 @@ import { downloadAudio } from '../process-steps/02-download-audio'
 import { runTranscription } from '../process-steps/03-run-transcription'
 import { selectPrompts } from '../process-steps/04-select-prompt'
 import { runLLM } from '../process-steps/05-run-llm'
-import { filterRSSItems, saveAudio, saveInfo } from '../utils/validate-option'
+import { filterRSSItems, saveAudio, saveInfo, parser } from '../utils/validate-option'
 import { l, err, logSeparator, logInitialFunctionCall, logRSSProcessingStatus } from '../utils/logging'
-import { parser } from '../utils/globals/process'
 import { retryRSSFetch } from '../utils/retry'
 
 import type { ProcessingOptions, RSSItem } from '../utils/types/process'
