+++ src/transcription/deepgram.ts
@@ -10,9 +10,8 @@
 
 import { readFile } from 'node:fs/promises'
 import { env } from 'node:process'
-import { l, err, logTranscriptionCost } from '../utils/logging'
-import { formatDeepgramTranscript } from './format-transcript'
-import { DEEPGRAM_MODELS } from '../utils/globals/transcription'
+import { l, err } from '../utils/logging'
+import { DEEPGRAM_MODELS, logTranscriptionCost, formatDeepgramTranscript } from '../utils/transcription-utils'
 import type { ProcessingOptions } from '../utils/types/process'
 import type { DeepgramResponse, DeepgramModelType } from '../utils/types/transcription'
